Small rendering to work buffer refactor for unified gsplats
